Added connectivity
