+++ cloudshell/networking/juniper/junos/junos_bootstrap.py
@@ -1,0 +1,3 @@
+from cloudshell.shell.core.driver_bootstrap import DriverBootstrap
+class JunosBootstrap(DriverBootstrap):
+    pass
