+++ cloudshell/networking/juniper/junos/resource_driver/junos_resource_driver.py
@@ -1,29 +1,50 @@
-from cloudshell.shell.core.driver_builder_wrapper import DriverFunction
-from cloudshell.networking.resource_driver.networking_generic_resource_driver import networking_generic_resource_driver
-import cloudshell.networking.juniper.junos
+from cloudshell.shell.core.resource_driver_interface import ResourceDriverInterface
+from cloudshell.networking.networking_resource_driver_interface import NetworkingResourceDriverInterface
+from cloudshell.networking.juniper.junos.junos_bootstrap import JunosBootstrap
+from cloudshell.networking.juniper.autoload.juniper_snmp_autoload_70 import JuniperSnmpAutoload70
+from cloudshell.shell.core.context_utils import context_from_args
+from cloudshell.networking.juniper.junos.handler.juniper_junos_connectivity_operations import \
+    JuniperJunosConnectivityOperations
 
-class junos_resource_driver(networking_generic_resource_driver):
-    ATTRIBUTE_MATRIX = {"resource": ["ResourceAddress", "User", "Password", "Enable Password", "Console Server IP Address",
-                                   "Console User", "Console Password", "Console Port", "CLI Connection Type",
-                                   "SNMP Version", "SNMP Read Community", "SNMP V3 User", "SNMP V3 Password",
-                                   "SNMP V3 Private Key"]}
-
-    @DriverFunction(extraMatrixRows=networking_generic_resource_driver.REQUIRED_RESORCE_ATTRIBUTES)
-    def Init(self, matrixJSON):
-        self.handler_name = 'JUNOS'
-        networking_generic_resource_driver.Init(self, matrixJSON)
-
-    @DriverFunction(alias='Shutdown', extraMatrixRows=networking_generic_resource_driver.REQUIRED_RESORCE_ATTRIBUTES)
-    def Shutdown(self, matrixJSON):
-        # self.__check_for_attributes_changes(matrixJSON)
-        self._check_for_attributes_changes(matrixJSON)
-        return self._resource_handler.shutdown()
-
-    @DriverFunction(alias='Send Config Command', extraMatrixRows=networking_generic_resource_driver.REQUIRED_RESORCE_ATTRIBUTES)
-    def SendCustomConfigCommand(self, matrixJSON, command):
-        self._check_for_attributes_changes(matrixJSON)
-        result_str = self._resource_handler.send_config_command(cmd=command)
-        return self._resource_handler.normalize_output(result_str)
+import cloudshell.networking.juniper.junos.junos_config as package_config
 
 
+class JunosResourceDriver(ResourceDriverInterface, NetworkingResourceDriverInterface):
+    def __init__(self):
+        bootstrap = JunosBootstrap()
+        bootstrap.add_config(package_config)
+        bootstrap.initialize()
 
+    def initialize(self, context):
+        pass
+
+    def cleanup(self):
+        pass
+
+    @context_from_args
+    def ApplyConnectivityChanges(self, context, request):
+        connectivity_handler = JuniperJunosConnectivityOperations()
+        connectivity_handler.apply_connectivity_changes(request)
+
+    def shutdown(self, context):
+        pass
+
+    @context_from_args
+    def get_inventory(self, context):
+        juniper_autoload = JuniperSnmpAutoload70()
+        return juniper_autoload.discover_snmp()
+
+    def save(self, context, folder_path, configuration_type):
+        pass
+
+    def send_custom_config_command(self, context, command):
+        pass
+
+    def send_custom_command(self, context, command):
+        pass
+
+    def update_firmware(self, context, remote_host, file_path):
+        pass
+
+    def restore(self, context, path, config_type, restore_method):
+        pass
